servodrive positioning

+++ Auxiliaries/delay.c
@@ -2,19 +2,19 @@
 
 #include <util/delay.h>
 
-void delay_s(long s) {
+void delay_s(int s) {
 	while(s--) {
 		_delay_ms(1000.0);
 	}
 }
 
-void delay_ms(long ms) {
+void delay_ms(int ms) {
 	while(ms--) {
 		_delay_ms(1.0);
 	}
 }
 
-void delay_us(long us) {
+void delay_us(int us) {
 	while(us--) {
 		_delay_us(1.0);
 	}
